Finalising Recompiler

+++ src/suyu_cmd/emu_window/emu_window_sdl2.h
@@ -83,7 +83,7 @@ protected:
     SDL_Window* render_window{};
 
     /// Keeps track of how often to update the title bar during gameplay
-    u32 last_time = 0;
+    u64 last_time = 0;
 
     /// Input subsystem to use with this window.
     InputCommon::InputSubsystem* input_subsystem;
